fix(api)!: map error codes for every method, and type credits as f64
Both changes are fixes for problems the new wiremock tests surfaced, so the
tests that documented them are updated here rather than left asserting the old
behaviour.

BREAKING: Organization::credits() returns f64, not i64.

The server declares latest_credit as Go float64. Deserializing into i64 meant an
organization holding a fractional credit failed org.get outright rather than
degrading. Go emits whole floats without a decimal point, which is why integral
values always round-tripped and hid it. The sibling UsageSummary already used
f64. CLI output is unchanged for whole numbers -- Rust prints 250.0f64 as "250"
-- and fractional credits now display instead of erroring.

BREAKING: JSON-RPC error codes are mapped to typed errors for every method.

map_rpc_error was invoked per call site and only three methods opted in:
job.run, job.stop and job.list. The other eighty-two returned a bare RpcError,
including auth.verify_token -- the first call every credentialled entry point
makes, where distinguishing an expired token from any other failure meant
matching the numeric code by hand.

(The previous commit's message said seven methods opted in. That was wrong: four
of the call sites I counted were inside mod tests_map_rpc_error, exercising the
function rather than using it. Three is the real number.)

The mapping now lives in process_rpc_response, through which every JSON-RPC
response passes, so it cannot be forgotten at a call site. The HTTP-status paths
in rpc_download and post_multipart route through it too, so an HTTP 403 and a
JSON-RPC 403 reach the caller as the same variant; that also subsumes two
hand-written 413 branches which produced exactly what map_rpc_error produces.

job.stop keeps its local call. Code 101 means task-not-found, and the typed
variant needs the task id, which rpc does not have -- so the code-only mappings
are applied centrally and the id-dependent one stays where the id exists. A
regression test pins that.

Callers matching RpcError(401 | 403, _) must now match PermissionDenied(method),
and RpcError(413, _) must match PayloadTooLarge. Other codes are unaffected.

+++ crates/edgefirst-client/src/api.rs
@@ -193,8 +193,14 @@ typeid!(
 pub struct Organization {
     id: OrganizationID,
     name: String,
+    /// `f64` to match the server, which declares `latest_credit` as Go
+    /// `float64`. This was `i64` until a wiremock test built from the server's
+    /// shape showed that a fractional credit fails to deserialize outright,
+    /// taking the whole `org.get` call with it. Go emits whole floats without a
+    /// decimal point, which is why integral values had always round-tripped and
+    /// hid the problem.
     #[serde(rename = "latest_credit")]
-    credits: i64,
+    credits: f64,
 }
 
 impl Display for Organization {
@@ -212,7 +218,7 @@ impl Organization {
         &self.name
     }
 
-    pub fn credits(&self) -> i64 {
+    pub fn credits(&self) -> f64 {
         self.credits
     }
 }
